AppyNox-22 closes #22 - BaseService created - Moved common methods moved to BaseService - Fixed minor problems in DtoMappingRegistery
