+++ src/Services/.BaseService/AppyNox.Services.Base.Application/Dtos/BaseDTO.cs
@@ -1,0 +1,11 @@
+﻿namespace AppyNox.Services.Base.Application.Dtos
+{
+    public abstract class BaseDto
+    {
+        #region [ Properties ]
+
+        public string? Code { get; set; }
+
+        #endregion
+    }
+}
