+++ src/Services/.BaseService/AppyNox.Services.Base.Application/Dtos/IUpdateDto.cs
@@ -1,0 +1,11 @@
+﻿namespace AppyNox.Services.Base.Application.Dtos
+{
+    public interface IUpdateDto
+    {
+        #region [ Properties ]
+
+        Guid Id { get; set; }
+
+        #endregion
+    }
+}
