+++ src/Services/.BaseService/AppyNox.Services.Base.Domain/AppyNox.Services.Base.Domain.csproj
@@ -6,8 +6,4 @@
     <Nullable>enable</Nullable>
   </PropertyGroup>
 
-  <ItemGroup>
-    <ProjectReference Include="..\..\.BaseService\AppyNox.Services.Base.Domain\AppyNox.Services.Base.Domain.csproj" />
-  </ItemGroup>
-
 </Project>
