+++ src/Services/.BaseService/AppyNox.Services.Base.Domain/Common/DtoLevelMappingTypes.cs
@@ -1,0 +1,17 @@
+﻿using System;
+using System.Collections.Generic;
+using System.Linq;
+using System.Text;
+using System.Threading.Tasks;
+
+namespace AppyNox.Services.Base.Domain.Common
+{
+    public enum DtoLevelMappingTypes
+    {
+        DataAccess,
+
+        Create,
+
+        Update
+    }
+}
