+++ src/Services/.BaseService/AppyNox.Services.Base.Domain/Interfaces/IEntityWithGuid.cs
@@ -1,0 +1,11 @@
+﻿namespace AppyNox.Services.Base.Domain.Interfaces
+{
+    public interface IEntityWithGuid
+    {
+        #region [ Properties ]
+
+        Guid Id { get; set; }
+
+        #endregion
+    }
+}
