+++ src/Services/.BaseService/AppyNox.Services.Base.Infrastructure/AppyNox.Services.Base.Infrastructure.csproj
@@ -1,0 +1,19 @@
+﻿<Project Sdk="Microsoft.NET.Sdk">
+
+	<PropertyGroup>
+		<TargetFramework>net8.0</TargetFramework>
+		<ImplicitUsings>enable</ImplicitUsings>
+		<Nullable>enable</Nullable>
+	</PropertyGroup>
+
+	<ItemGroup>
+		<PackageReference Include="Microsoft.EntityFrameworkCore.DynamicLinq" Version="7.3.5" />
+		<PackageReference Include="LinqKit.Microsoft.EntityFrameworkCore" Version="7.1.4" />
+		<PackageReference Include="Microsoft.EntityFrameworkCore" Version="7.0.11" />
+	</ItemGroup>
+
+	<ItemGroup>
+		<ProjectReference Include="..\AppyNox.Services.Base.Domain\AppyNox.Services.Base.Domain.csproj" />
+	</ItemGroup>
+
+</Project>
