+++ src/Services/.BaseService/AppyNox.Services.Base.Infrastructure/Interfaces/IUnitOfWorkBase.cs
@@ -1,0 +1,17 @@
+﻿namespace AppyNox.Services.Base.Infrastructure.Interfaces
+{
+    public interface IUnitOfWorkBase : IDisposable
+    {
+        #region [ Public Methods ]
+
+        void BeginTransaction();
+
+        void Commit();
+
+        void Rollback();
+
+        Task<int> SaveChangesAsync();
+
+        #endregion
+    }
+}
