--- src/Services/CouponService/AppyNox.Services.Coupon.Application/DTOs/BaseDTO.cs
@@ -1,11 +1,0 @@
-﻿namespace AppyNox.Services.Coupon.Application.Dtos
-{
-    public abstract class BaseDto
-    {
-        #region [ Properties ]
-
-        public string? Code { get; set; }
-
-        #endregion
-    }
-}
