--- src/Services/CouponService/AppyNox.Services.Coupon.Application/DTOs/IUpdateDTO.cs
@@ -1,11 +1,0 @@
-﻿namespace AppyNox.Services.Coupon.Application.Dtos
-{
-    public interface IUpdateDto
-    {
-        #region [ Properties ]
-
-        Guid Id { get; set; }
-
-        #endregion
-    }
-}
