+++ src/Services/CouponService/AppyNox.Services.Coupon.Application/Dtos/CouponDetailDtos/Models/Basic/CouponDetailSimpleDto.cs
@@ -1,11 +1,10 @@
-﻿using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Text;
-using System.Threading.Tasks;
+﻿using AppyNox.Services.Base.Application.Dtos;
+using AppyNox.Services.Coupon.Application.Dtos.CouponDetailDtos.DetailLevel;
+
 
 namespace AppyNox.Services.Coupon.Application.Dtos.CouponDetailDtos.Models.Basic
 {
+    [CouponDetailDetailLevel(CouponDetailDataAccessDetailLevel.Simple)]
     public class CouponDetailSimpleDto : BaseDto
     {
         #region [ Properties ]
