+++ src/Services/CouponService/AppyNox.Services.Coupon.Application/Dtos/CouponDetailDtos/Models/Basic/CouponDetailSimpleUpdateDto.cs
@@ -8,5 +8,6 @@ namespace AppyNox.Services.Coupon.Application.Dtos.CouponDetailDtos.Models.Basic
 {
     internal class CouponDetailSimpleUpdateDto
     {
+        // mustang TODO Implement
     }
 }
