+++ src/Services/CouponService/AppyNox.Services.Coupon.Application/Dtos/CouponDtos/Models/Base/CouponSimpleUpdateDto.cs
@@ -1,4 +1,5 @@
-﻿using AppyNox.Services.Coupon.Application.Dtos.CouponDtos.DetailLevel;
+﻿using AppyNox.Services.Base.Application.Dtos;
+using AppyNox.Services.Coupon.Application.Dtos.CouponDtos.DetailLevel;
 
 namespace AppyNox.Services.Coupon.Application.Dtos.CouponDtos.Models.Base
 {
