--- src/Services/CouponService/AppyNox.Services.Coupon.Application/Validators/SharedRules/GenericValidationRules.cs
@@ -1,26 +1,0 @@
-﻿using FluentValidation;
-
-namespace AppyNox.Services.Coupon.Application.Validators.SharedRules
-{
-    public static class GenericValidationRules
-    {
-        #region [ Public Methods ]
-
-        public static IRuleBuilderOptions<T, string?> ValidateCode<T>(this IRuleBuilder<T, string?> ruleBuilder)
-        {
-            return ruleBuilder
-                .NotNull().NotEmpty().WithMessage("Code cannot be null")
-                .Length(5).WithMessage("Code should be 5 characters");
-        }
-
-        public static IRuleBuilderOptions<T, Guid> ValidateId<T>(this IRuleBuilder<T, Guid> ruleBuilder)
-        {
-            return ruleBuilder
-                .NotNull()
-                .NotEmpty()
-                .WithMessage("Id cannot be null");
-        }
-
-        #endregion
-    }
-}
