--- src/Services/CouponService/AppyNox.Services.Coupon.Domain/Common/CommonDtoLevelEnums.cs
@@ -1,22 +1,0 @@
-﻿using System;
-using System.Collections.Generic;
-using System.ComponentModel;
-using System.ComponentModel.DataAnnotations;
-using System.Linq;
-using System.Text;
-using System.Threading.Tasks;
-
-namespace AppyNox.Services.Coupon.Domain.Common
-{
-    public enum CommonDtoLevelEnums
-    {
-        [Display(Name = "None")]
-        None,
-
-        [Display(Name = "Simple")]
-        Simple,
-
-        [Display(Name = "IdOnly")]
-        IdOnly
-    }
-}
