--- src/Services/CouponService/AppyNox.Services.Coupon.Domain/Common/DtoLevelMappingTypes.cs
@@ -1,17 +1,0 @@
-﻿using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Text;
-using System.Threading.Tasks;
-
-namespace AppyNox.Services.Coupon.Domain.Common
-{
-    public enum DtoLevelMappingTypes
-    {
-        DataAccess,
-
-        Create,
-
-        Update
-    }
-}
