+++ src/Services/CouponService/AppyNox.Services.Coupon.Domain/Common/QueryParameters.cs
@@ -1,108 +1,27 @@
-﻿using System.ComponentModel.DataAnnotations;
-using AppyNox.Services.Coupon.Domain.Interfaces.Common;
+﻿using AppyNox.Services.Base.Domain.Common;
+using System;
+using System.Collections.Generic;
+using System.Linq;
+using System.Text;
+using System.Threading.Tasks;
 
 namespace AppyNox.Services.Coupon.Domain.Common
 {
-    public class QueryParameters : IQueryParameters
+    public class QueryParameters : QueryParametersBase
     {
-        #region [ Properties ]
-
-        private CommonDtoLevelEnums _commonDtoLevel = CommonDtoLevelEnums.None;
-
-        private string _detailLevel = string.Empty;
-
-        private DtoLevelMappingTypes _accessType = DtoLevelMappingTypes.DataAccess;
-
-        [Range(1, int.MaxValue, ErrorMessage = "Page number must be greater than 0.")]
-        public int PageNumber { get; set; } = 1;
-
-        [Range(1, 100, ErrorMessage = "Page size must be between 1 and 100.")]
-        public int PageSize { get; set; } = 10;
-
-        public CommonDtoLevelEnums CommonDtoLevel { get => _commonDtoLevel; }
-
-        public DtoLevelMappingTypes AccessType
-        {
-            get => _accessType;
-            internal set => _accessType = value;
-        }
-
-        /// <summary>
-        /// Used only for converting AccessType, dont use this variable in code.
-        /// </summary>
-        public string Access
-        {
-            get => string.Empty;
-            set
-            {
-                _accessType = value switch
-                {
-                    "DataAccess" => DtoLevelMappingTypes.DataAccess,
-                    "Update" => DtoLevelMappingTypes.Update,
-                    "Create" => DtoLevelMappingTypes.Create,
-                    _ => DtoLevelMappingTypes.DataAccess
-                };
-            }
-        }
-
-        public string DetailLevel
-        {
-            get => _detailLevel;
-            set
-            {
-                if (string.IsNullOrEmpty(value))
-                {
-                    _detailLevel = "Simple";
-                    _commonDtoLevel = CommonDtoLevelEnums.Simple;
-                }
-                else
-                {
-                    _detailLevel = value;
-                    _commonDtoLevel = value switch
-                    {
-                        "Simple" => CommonDtoLevelEnums.Simple,
-                        _ => CommonDtoLevelEnums.None
-                    };
-                }
-            }
-        }
-
-        #endregion
-
         #region [ Public Constructors ]
 
-        public QueryParameters()
+        public QueryParameters() : base()
         {
-            DetailLevel = string.Empty;
         }
 
-        public static QueryParameters CreateForIdOnly()
+        public QueryParameters(CommonDtoLevelEnums commonDtoLevel) : base(commonDtoLevel)
         {
-            return new QueryParameters(CommonDtoLevelEnums.IdOnly);
         }
 
-        #endregion
-
-        #region [ Private Constructors ]
-
-        private QueryParameters(CommonDtoLevelEnums commonDtoLevel) : this()
+        public new static QueryParametersBase CreateForIdOnly()
         {
-            switch (commonDtoLevel)
-            {
-                case CommonDtoLevelEnums.Simple:
-                    DetailLevel = "Simple";
-                    _commonDtoLevel = CommonDtoLevelEnums.Simple;
-                    break;
-
-                case CommonDtoLevelEnums.IdOnly:
-                    DetailLevel = "IdOnly";
-                    _commonDtoLevel = CommonDtoLevelEnums.IdOnly;
-                    break;
-
-                case CommonDtoLevelEnums.None:
-                default:
-                    throw new ArgumentException("CommonDtoLevel is not provided or set as none for QueryParameters.");
-            }
+            return new QueryParameters(CommonDtoLevelEnums.IdOnly) as QueryParametersBase;
         }
 
         #endregion
