--- src/Services/CouponService/AppyNox.Services.Coupon.Domain/Interfaces/IEntityWithGuid.cs
@@ -1,11 +1,0 @@
-﻿namespace AppyNox.Services.Coupon.Domain.Interfaces
-{
-    public interface IEntityWithGuid
-    {
-        #region [ Properties ]
-
-        Guid Id { get; set; }
-
-        #endregion
-    }
-}
