--- src/Services/CouponService/AppyNox.Services.Coupon.Domain/Interfaces/IUnitOfWork.cs
@@ -1,17 +1,0 @@
-﻿namespace AppyNox.Services.Coupon.Domain.Interfaces
-{
-    public interface IUnitOfWork : IDisposable
-    {
-        #region [ Public Methods ]
-
-        void BeginTransaction();
-
-        void Commit();
-
-        void Rollback();
-
-        Task<int> SaveChangesAsync();
-
-        #endregion
-    }
-}
